backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python==4.10.0.84
deepface==0.0.79
numpy
tensorflow
tf-keras

# Voice transcription dependencies
pyaudio==0.2.14
SpeechRecognition==3.10.4
openai-whisper==20231117
google-genai
python-dotenv==1.0.0
pyttsx3==2.90

# Facial sentiment analysis dependencies
fer==22.5.1
keras==2.15.0
